Switch to esModules and Webpack
