+++ README.md
@@ -1,2 +1,6 @@
 # National Parks Data Visualization
 TBD
+
+TODO: break D3 into smaller bundles
+TODO: update webpack config
+TODO: compress/gzip images?
