+++ package.json
@@ -2,15 +2,23 @@
   "name": "national-parks",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "reload --browser --dir public"
+    "build": "webpack && cp -r public/* dist",
+    "start": "webpack serve --open"
   },
   "keywords": [],
-  "author": "",
+  "author": "brouse12",
   "license": "ISC",
   "devDependencies": {
-    "reload": "^3.2.0"
+    "csv-loader": "^3.0.5",
+    "webpack": "^5.74.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.11.1"
+  },
+  "dependencies": {
+    "@googlemaps/js-api-loader": "^1.14.3",
+    "d3": "^7.6.1"
   }
 }
