--- public/data/index.html
@@ -1,13 +1,0 @@
-<html xmlns="http://www.w3.org/1999/xhtml" >
-<head>
-    <title>National Parks</title>
-    <link rel="shortcut icon" type="image/ico" href="../favicon.ico"/>
-    <meta charset="utf-8">
-    <meta name="description" content="Learn more about USA national parks">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-</head>
-<body>
-  <h1>404 Not Found</h1>
-  <p>Sorry, we could not find this page! Instead, why don't you try to find a National Park?</p>
-</body>
-</html>
