+++ public/index.html
@@ -2,16 +2,15 @@
 <html>
     <head>
         <title>National Parks</title>
-        <link rel="shortcut icon" type="image/ico" href="./favicon.ico"/>
+        <link rel="shortcut icon" type="image/ico" href="favicon.ico"/>
         <meta charset="utf-8">
         <meta name="description" content="Learn more about USA national parks">
         <meta name="viewport" content="width=device-width, initial-scale=1">
         <link rel="stylesheet" href="styles.css">
-        <script src="https://d3js.org/d3.v7.min.js" defer></script>
-        <script src="parallax.js" defer></script>
+        <script src="bundle.js"></script>
     </head>
     <body>
-        <div class="loading-background">
+        <div class="loading-background" id="loading-background">
             <div class="loading-indicator">
                 <div></div><div></div><div></div><div></div><div></div><div></div>
                 <div></div><div></div><div></div><div></div><div></div><div></div>
