+++ public/styles.css
@@ -47,10 +47,6 @@ body {
   }
 }
 
-
-
-
-
 /* ------------------- Loading Screen ----------------- */
 
 .loading-background {
